Add manual printing in program controller

+++ LibraryApp/ProgramController.cs
@@ -25,41 +25,52 @@ namespace LibraryApp
 
         public void Start(string[] args)
         {
-            string action = args[0];
-            string result = "";
-            IList<Book> books = new List<Book>();
-
-            switch (action)
+            if (args.Length != 0)
             {
-                case "add":
-                    result = _bookService.AddNewBook(args);
-                    break;
-                case "list-all":
-                    books = _bookService.GetAListOfAllBooks();
-                    break;
-                case "delete":
-                    result = _bookService.DeleteBook(args);
-                    break;
-                case "filter":
-                    books = _filteringService.FilterBooks(args[1]);
-                    break;
-                case "take":
-                    result = _borrowingService.TakeABook(args);
-                    break;
-                case "return":
-                    result = _returnService.ReturnBook(args);
-                    break;
-                default:
-                    break;
-            }
+                string action = args[0];
+                string result = "";
+                IList<Book> books = null;
 
-            if (!string.IsNullOrWhiteSpace(result))
-            {
-                _consoleWriter.PrintResultMessage(result);
+                switch (action)
+                {
+                    case "add":
+                        result = _bookService.AddNewBook(args);
+                        break;
+                    case "list-all":
+                        books = _bookService.GetAListOfAllBooks();
+                        break;
+                    case "delete":
+                        result = _bookService.DeleteBook(args);
+                        break;
+                    case "filter":
+                        books = _filteringService.FilterBooks(args[1]);
+                        break;
+                    case "take":
+                        result = _borrowingService.TakeABook(args);
+                        break;
+                    case "return":
+                        result = _returnService.ReturnBook(args);
+                        break;
+                    case "help":
+                        _consoleWriter.PrintManual();
+                        break;
+                    default:
+                        _consoleWriter.PrintManual();
+                        break;
+                }
+
+                if (!string.IsNullOrWhiteSpace(result))
+                {
+                    _consoleWriter.PrintResultMessage(result);
+                }
+                else if (books != null)
+                {
+                    _consoleWriter.PrintBooks(books);
+                }
             }
             else
             {
-                _consoleWriter.PrintBooks(books);
+                _consoleWriter.PrintManual();
             }
         }
     }
